添加了搜索服务

+++ b2b-db/src/main/java/com/b2b/mall/db/mapper/ItemDescMapper.java
@@ -13,7 +13,7 @@ public interface ItemDescMapper {
 
     int updateByPrimaryKeySelective(ItemDesc record);
 
-    ItemDesc selectItemDescByPrimaryKey(Integer itemId);
+    ItemDesc selectItemDescByPrimaryKey(Long itemId);
 
     int updateByPrimaryKeyWithBLOBs(ItemDesc record);
